Add client/server authentication support
- By default servers are authenticated unless an option is passed to be unauthenticated
	- Authenticated servers have all methods other than Authenticate guarded by auth checks
	- Successful authentication returns a JWT to use for future requests
	- Note: WebRTC connections assume authenticated signaling and bypass auth checks
- Making a new server allows registering auth handlers at the same time for specific credential types
	- See rpc/examples/echo (make run_server_auth)
	- Authorization is left up to the library user
- Dial now accepts variadic options
- NewServer now accepts variadic options

+++ etc/setup_cert.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 
-mkdir -p $HOME/.minica
-cd $HOME/.minica
+mkdir -p test_keys
+cd test_keys
+if [[ -f "localhost/cert.pem" && -f "localhost/key.pem" ]]; then
+	exit 0;
+fi
 go install github.com/jsha/minica@latest
 minica --domains localhost
-#TODO(erd): support linux
+#TODO(https://github.com/viamrobotics/goutils/issues/9): support linux
 sudo security add-trusted-cert -d -r trustRoot -k /Library/Keychains/System.keychain minica.pem
